Rename file of common extractors to 'common'
