+++ classification/feature_extraction/FeatureExtractionPipeline.py
@@ -1,16 +1,16 @@
 from itertools import chain
+
 from github.Repository import Repository
 
 from classification.Feature import Feature
-from classification.feature_extraction.CommonFeatureExtractors import ContributorsExtractor, BranchExtractor, \
-    ForkExtractor, StarExtractor, CommitNumberExtractor
+from classification.feature_extraction import common
 
 FEATURE_EXTRACTORS = [
-    BranchExtractor,
-    CommitNumberExtractor,
-    ContributorsExtractor,
-    ForkExtractor,
-    StarExtractor,
+    common.BranchExtractor,
+    common.CommitNumberExtractor,
+    common.ContributorsExtractor,
+    common.ForkExtractor,
+    common.StarExtractor,
 ]
 
 
@@ -19,5 +19,4 @@ class FeatureExtractionPipeline:
         self._repo = repo
 
     def extract_features(self) -> [Feature]:
-		return list(chain.from_iterable(
-            (extractor(self._repo).extract_features() for extractor in FEATURE_EXTRACTORS)))
+        return list(chain.from_iterable((extractor(self._repo).extract_features() for extractor in FEATURE_EXTRACTORS)))
