+++ classification/tests/test_common_extractors.py
@@ -1,6 +1,6 @@
 from unittest import TestCase
 
-from classification.feature_extraction.CommonFeatureExtractors import TotalFilesExtractor, CommitNumberExtractor
+from classification.feature_extraction.common import TotalFilesExtractor, CommitNumberExtractor
 from classification.tests.GithubInitializerForTest import GithubInitializerForTest
 
 # TODO probably create a own repository with individual content to test various setups
